code_release/src/pyproject.toml
[project]
name = "transformer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython==1.85",
    "huggingface-hub==0.34.1",
    "matplotlib==3.10.6",
    "mpmath==1.3.0",
    "multimolecule==0.0.8",
    "numpy==2.3.1",
    "openpyxl==3.1.5",
    "optuna==4.5.0",
    "pandas==2.3.1",
    "plotly==6.3.0",
    "rna-fm==0.2.2",
    "scienceplots==2.1.1",
    "scikit-learn==1.7.1",
    "scipy==1.16.0",
    "seaborn==0.13.2",
    "shap==0.48.0",
    "tensorboard==2.20.0",
    "torch==2.7.1",
    "torcheval==0.0.7",
    "tqdm==4.67.1",
    "transformers==4.54.0",
    "viennarna==2.7.0",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform == 'linux' or sys_platform == 'win64'" },
]

[[tool.uv.index]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
default = true
